#71 state情報をContainer.jsへリフトアップ（途中）
