+++ src/App.js
@@ -1,19 +1,15 @@
-import { Route, Routes } from "react-router-dom";
+// import { Route, Routes } from "react-router-dom";
 
 import React from "react";
-import Home from "./pages/Home.js";
-import Settings from "./pages/Settings";
-import GameDisplayPage from "./pages/GameDisplayPage";
-import Header from "./components/Header";
+// import Home from "./pages/Home.js";
+// import Settings from "./pages/Settings";
+// import GameDisplayPage from "./pages/GameDisplayPage";
+// import Header from "./components/Header";
+import Container from "./Container";
 function App() {
   return (
     <div>
-      <Header />
-      <Routes>
-        <Route path="/" element={<Home />}></Route>
-        <Route path="/settings" element={<Settings />}></Route>
-        <Route path="/game" element={<GameDisplayPage />}></Route>
-      </Routes>
+      <Container />
     </div>
   );
 }
